fix(settings): update code limits and add new paths for session management fix(async_launcher): enhance log path allocation with atomic file creation feat(handoff_render): add rendering utilities for agent chains and handoff events refactor(handoff_read): clean up imports and remove unused functions refactor(session_registry): simplify type checking and logger imports
