+++ src/vibe3/commands/handoff_read.py
@@ -10,8 +10,12 @@ from loguru import logger
 from vibe3.agents.backends.codeagent import CodeagentBackend
 from vibe3.clients.sqlite_client import SQLiteClient
 from vibe3.commands.common import trace_scope
+from vibe3.commands.handoff_render import (
+    _render_agent_chain,
+    _render_handoff_events,
+    _render_updates_log,
+)
 from vibe3.environment.session_registry import SessionRegistryService
-from vibe3.models.flow import FlowEvent, FlowState
 from vibe3.services.flow_service import FlowService
 from vibe3.services.handoff_service import HandoffService
 from vibe3.ui.console import console
@@ -22,8 +26,6 @@ from vibe3.ui.handoff_ui import (
 )
 from vibe3.utils.git_helpers import get_branch_handoff_dir
 from vibe3.utils.issue_branch_resolver import resolve_issue_branch_input
-
-UPDATE_LOG_MESSAGE_PREVIEW_LIMIT = 80
 
 
 def _get_live_sessions_for_branch(store: SQLiteClient, branch: str) -> list[dict]:
@@ -42,164 +44,6 @@ def _get_live_sessions_for_branch(store: SQLiteClient, branch: str) -> list[dict
     backend = CodeagentBackend()
     registry = SessionRegistryService(store=store, backend=backend)
     return registry.get_truly_live_sessions_for_branch(branch)
-
-
-def _resolve_ref_path(val: str | None, _worktree_root: str | None = None) -> str:
-    """Resolve ref value to absolute path if needed, but prefer relative.
-
-    For file paths, returns the original relative path if it's already
-    relative. For absolute paths, keeps them absolute (e.g., URLs).
-
-    Args:
-        val: The reference value (could be relative path, absolute path,
-            URL, etc.)
-        _worktree_root: The worktree root path (unused, kept for
-            signature compatibility)
-
-    Returns:
-        Display-friendly path (relative preferred for file refs)
-    """
-    # Note: _worktree_root intentionally unused - display relative paths
-    _ = _worktree_root  # Mark as intentionally unused for type checker
-
-    if not val or Path(val).is_absolute() or val.startswith("("):
-        return val or ""
-    # For relative file paths, return as-is (cleaner display)
-    return val
-
-
-def _render_agent_chain(
-    state: FlowState,
-    store: SQLiteClient,
-    live_sessions: list[dict] | None = None,
-    worktree_root: str | None = None,
-) -> None:
-    """Render agent chain with issue URL fallback for spec_ref."""
-    console.print("[bold]Agent Chain[/]")
-
-    # Check if this is an issue-based flow
-    from vibe3.clients.github_client import GitHubClient
-    from vibe3.config.settings import VibeConfig
-    from vibe3.services.issue_flow_service import IssueFlowService
-
-    issue_service = IssueFlowService(store=store)
-    issue_number = issue_service.parse_issue_number_any(state.branch)
-
-    for label, actor_label in [
-        ("spec_ref", "planner_actor"),
-        ("plan_ref", "planner_actor"),
-        ("report_ref", "executor_actor"),
-        ("audit_ref", "reviewer_actor"),
-    ]:
-        val = getattr(state, label, None)
-        actor = getattr(state, actor_label, None) or ""
-        actor_str = f"  [dim]{actor}[/]" if actor else ""
-
-        # Special handling for spec_ref: show issue URL if None but issue flow
-        if label == "spec_ref" and not val and issue_number is not None:
-            # Try to get issue URL from cache or GitHub API
-            cache = store.get_flow_context_cache(state.branch)
-            issue_url = None
-            effective_issue_number = issue_number
-
-            # Try cache first for issue number
-            if cache and cache.get("task_issue_number"):
-                effective_issue_number = cache["task_issue_number"]
-
-            # Build GitHub issue URL
-            config = VibeConfig.get_defaults()
-            repo = config.orchestra.repo if config.orchestra.repo else None
-
-            if not repo:
-                # Fallback: try to get repo from gh CLI (current repo)
-                try:
-                    import subprocess
-
-                    result = subprocess.run(
-                        ["gh", "repo", "view", "--json", "nameWithOwner"],
-                        capture_output=True,
-                        text=True,
-                        check=True,
-                    )
-                    import json
-
-                    data = json.loads(result.stdout)
-                    repo = data.get("nameWithOwner")
-                except Exception:
-                    pass
-
-            if repo:
-                # repo format: "owner/repo"
-                issue_url = f"https://github.com/{repo}/issues/{effective_issue_number}"
-            else:
-                # Last fallback: try to get from GitHub API
-                try:
-                    gh = GitHubClient()
-                    issue_data = gh.view_issue(effective_issue_number)
-                    if isinstance(issue_data, dict):
-                        issue_url = issue_data.get("html_url")
-                except Exception:
-                    pass
-
-            if issue_url:
-                console.print(f"  [dim]{label}[/]{actor_str}")
-                console.print(
-                    f"    [link={issue_url}]{issue_url}[/]",
-                    no_wrap=True,
-                    overflow="ellipsis",
-                )
-            else:
-                # Fallback: show issue number with pending status
-                console.print(f"  [dim]{label}[/]{actor_str}")
-                console.print(
-                    f"    [dim]#{effective_issue_number} (issue flow, pending spec)[/]"
-                )
-        elif label == "spec_ref" and not val and issue_number is None:
-            # Non-issue flow with no spec_ref
-            console.print(f"  [dim]{label}[/]  [dim](pending)[/]")
-        else:
-            # Normal ref display (plan_ref, report_ref, audit_ref)
-            display_val: str = _resolve_ref_path(val, worktree_root)
-            if display_val:
-                label_line = f"  [dim]{label}[/]{actor_str}"
-                console.print(label_line)
-                console.print(f"    {display_val}", no_wrap=True, overflow="ellipsis")
-            else:
-                console.print(f"  [dim]{label}[/]  [dim](pending)[/]")
-    console.print()
-
-    # Show live registry sessions (registry is the source of truth)
-    if live_sessions:
-        console.print("[bold]Live Sessions (registry)[/]")
-        for sess in live_sessions:
-            role = sess.get("role", "")
-            name = sess.get("session_name", "")
-            status = sess.get("status", "")
-            console.print(f"  [green]{role}[/]  {name}  [dim]{status}[/]")
-        console.print()
-
-
-def _render_handoff_events(events: list[FlowEvent]) -> None:
-    if not events:
-        console.print("[dim]  no handoff events[/]")
-        return
-
-    for event in reversed(events):
-        time_str = event.created_at[:19].replace("T", " ")
-        console.print(
-            f"[dim]{time_str}[/]  [magenta]{event.event_type}[/]  [dim]{event.actor}[/]"
-        )
-        if event.detail:
-            console.print(f"  {event.detail}")
-        if event.refs:
-            files = event.refs.get("files") if isinstance(event.refs, dict) else None
-            if files and isinstance(files, list):
-                for f in files:
-                    console.print(f"  [dim]- {f}[/]")
-            ref = event.refs.get("ref") if isinstance(event.refs, dict) else None
-            if ref:
-                console.print(f"  [dim]- {ref}[/]")
-        console.print()
 
 
 def _parse_updates_section(content: str) -> list[dict[str, str]]:
@@ -235,46 +79,6 @@ def _parse_updates_section(content: str) -> list[dict[str, str]]:
             else:
                 updates[-1]["message"] = line
     return updates
-
-
-def _preview_update_message(message: str, truncate: bool) -> str:
-    """Preview update message with intelligent truncation.
-
-    If message contains file paths (contains '/'), avoid truncating mid-path.
-    Otherwise, use standard character limit truncation.
-    """
-    if not truncate or len(message) <= UPDATE_LOG_MESSAGE_PREVIEW_LIMIT:
-        return message
-
-    # Check if message looks like it contains a file path
-    if "/" in message:
-        # For file path messages, don't truncate - let console handle overflow
-        return message
-
-    # For non-path messages, use standard truncation
-    return message[:UPDATE_LOG_MESSAGE_PREVIEW_LIMIT] + "..."
-
-
-def _render_updates_log(updates: list[dict[str, str]], truncate: bool = True) -> None:
-    """Render updates in log format."""
-    if not updates:
-        console.print("[dim]  no updates yet[/]")
-        return
-
-    kind_colors = {"finding": "yellow", "blocker": "red", "next": "blue", "note": "dim"}
-    for update in updates:
-        timestamp = update["timestamp"]
-        actor = update["actor"]
-        kind = update["kind"]
-        message = update["message"]
-        kind_color = kind_colors.get(kind, "dim")
-        time_str = timestamp[:19].replace("T", " ")
-        console.print(f"[dim]{time_str}[/]  [{kind_color}]{kind}[/]  [dim]{actor}[/]")
-        if message:
-            for msg_line in _preview_update_message(message, truncate).split("\n"):
-                # Use overflow='ellipsis' for clean path truncation if needed
-                console.print(f"  {msg_line}", overflow="ellipsis")
-        console.print()
 
 
 def list_handoffs(
